% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tscgm.ar2.r
\name{plot.tscgm.ar2}
\alias{plot.tscgm.ar2}
\title{Plot method for tscgm.ar2 objects}
\usage{
\method{plot}{tscgm.ar2}(x, mat = c("precision", "autoregression1", "autoregression2"), ...)
}
\arguments{
\item{x}{An object of class \code{tscgm.ar2}.}

\item{mat}{Character. Which matrix to plot. Choices are
\code{"precision"}, \code{"autoregression1"}, or \code{"autoregression2"}.}

\item{...}{Additional plotting arguments passed to the plotting function.}
}
\description{
Plots matrices from \code{tscgm.ar2} objects. You can choose to plot the precision matrix,
the first-order autoregressive matrix, or the second-order autoregressive matrix.
}
