% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tscgm.R
\name{plot.tscgm}
\alias{plot.tscgm}
\title{Plot method for tscgm objects}
\usage{
\method{plot}{tscgm}(x, mat = c("precision", "autoregression"), ...)
}
\arguments{
\item{x}{An object of class \code{tscgm}.}

\item{mat}{Character. Which matrix to plot. Choices are \code{"precision"} or \code{"autoregression"}.}

\item{...}{Additional plotting arguments passed to the plotting function.}
}
\description{
Plots matrices from tscgm objects. You can choose to plot the precision matrix
or the autoregressive coefficient matrix.
}
