\name{print.tscgm}
\alias{print.tscgm}
\title{
Print summary - S3 Method for Class 'sparse.tscgm'.
}
\description{
  \code{print} is a generic function that prints output summaries of fitted
  models in the SparseTSCGM package. 
}
\usage{
 \method{print}{tscgm}(x, ...)
}
\arguments{
  \item{x}{
    an object of class \code{sparse.tscgm}
  }
  \item{...}{
    other arguments passed to print.
  }
}
\details{
The \code{print.tscgm} function summarizes and prints results of 
 the fitted model.}
\value{
  Prints summary of results:
}
\author{
Fentaw Abegaz and Ernst Wit
}
\seealso{
  \code{\link{sparse.tscgm}}
}
