% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tscgm.R
\name{print.tscgm}
\alias{print.tscgm}
\title{Print method for tscgm objects}
\usage{
\method{print}{tscgm}(x, ...)
}
\arguments{
\item{x}{An object of class tscgm}

\item{...}{Further arguments}
}
\description{
Print method for tscgm objects
}
