% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.tscgm.R
\name{summary.tscgm}
\alias{summary.tscgm}
\title{Summary method for tscgm objects}
\usage{
\method{summary}{tscgm}(object, ...)
}
\arguments{
\item{object}{An object of class tscgm}

\item{...}{Further arguments}
}
\description{
Summary method for tscgm objects
}
